requirements.txt
pandas
pillow
plotly
streamlit